requirements.txt
asgiref==3.4.1
bleach==4.1.0
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.7
cryptography==35.0.0
defusedxml==0.7.1
Django==3.2.9
django-cors-headers==3.10.0
django-crispy-forms==1.13.0
django-extensions==3.1.5
django-summernote==0.8.20.0
django-taggit==2.0.0
djangorestframework==3.12.4
idna==3.3
oauthlib==3.1.1
packaging==21.3
Pillow==8.4.0
psycopg2-binary==2.9.2
pycparser==2.21
PyJWT==2.3.0
pyparsing==3.0.6
python3-openid==3.2.0
pytz==2021.3
requests==2.26.0
requests-oauthlib==1.3.0
six==1.16.0
social-auth-app-django==5.0.0
social-auth-core==4.1.0
sqlparse==0.4.2
urllib3==1.26.7
webencodings==0.5.1
